+++ p2panda-rs/src/document/document.rs
@@ -259,7 +259,6 @@ mod tests {
     use crate::identity::KeyPair;
     use crate::operation::{Operation, OperationId, OperationValue, OperationWithMeta};
     use crate::schema::SchemaId;
-    use crate::test_utils::constants::DEFAULT_SCHEMA_HASH;
     use crate::test_utils::fixtures::{
         create_operation, delete_operation, fields, random_key_pair, schema, update_operation,
     };
@@ -452,8 +451,8 @@ mod tests {
         let expected_graph_tips: Vec<OperationId> = vec![penguin_entry_3_hash.clone().into()];
         let expected_op_order = vec![
             panda_1.clone(),
-            panda_2.clone(),
             penguin_1.clone(),
+            panda_2.clone(),
             penguin_2.clone(),
             penguin_3.clone(),
         ];
@@ -524,7 +523,7 @@ mod tests {
     }
 
     #[rstest]
-    fn doc_test() {
+    fn doc_test(schema: SchemaId) {
         let polar = Client::new(
             "polar".to_string(),
             KeyPair::from_private_key_str(
@@ -539,7 +538,7 @@ mod tests {
             )
             .unwrap(),
         );
-        let schema = SchemaId::new(DEFAULT_SCHEMA_HASH).unwrap();
+
         let mut node = Node::new();
         let (polar_entry_1_hash, _) = send_to_node(
             &mut node,
@@ -573,7 +572,10 @@ mod tests {
             &update_operation(
                 schema.clone(),
                 vec![polar_entry_1_hash.clone().into()],
-                operation_fields(vec![("name", OperationValue::Text("🐼 Cafe!".to_string()))]),
+                operation_fields(vec![(
+                    "name",
+                    OperationValue::Text("🐼 Cafe!!".to_string()),
+                )]),
             ),
         )
         .unwrap();
@@ -688,7 +690,7 @@ mod tests {
         // Here we see that "🐼 Cafe!" won the conflict, meaning it was applied after "ʕ •ᴥ•ʔ Cafe!".
         assert_eq!(
             document_view.get("name").unwrap(),
-            &OperationValue::Text("🐼 Cafe!".into())
+            &OperationValue::Text("🐼 Cafe!!".into())
         );
         assert_eq!(
             document_view.get("owner").unwrap(),
@@ -725,7 +727,7 @@ mod tests {
 
         assert_eq!(
             document_view.get("name").unwrap(),
-            &OperationValue::Text("🐼 Cafe!".into())
+            &OperationValue::Text("🐼 Cafe!!".into())
         );
         assert_eq!(
             document_view.get("owner").unwrap(),