+++ p2panda-rs/src/lib.rs
@@ -25,7 +25,7 @@
 //! # let profile_schema_view_id = OperationId::from(
 //! #     Hash::new_from_bytes(vec![1, 2, 3])?
 //! # );
-//! # let profile_schema: SchemaId = profile_schema_view_id.into();
+//! # let profile_schema = SchemaId::new_application("profile", &profile_schema_view_id.into());
 //! // Generate new Ed25519 key pair
 //! let key_pair = KeyPair::new();
 //!