+++ CHANGELOG.md
@@ -38,7 +38,8 @@ Highlights are marked with a pancake 🥞
 - Improve error reporting when adding operation fields [#262](https://github.com/p2panda/p2panda/issues/262)] `rs` `js`
 - Update mock node API [#286](https://github.com/p2panda/p2panda/issues/286) `rs`
 - Refactored graph module to be generic over graph node keys and other graph improvements [#289](https://github.com/p2panda/p2panda/issues/289) `rs`
-- Require sorted serialisation of document view ids [#284](https://github.com/p2panda/p2panda/pull/284) `js`
+- Require sorted serialisation of document view ids [#284](https://github.com/p2panda/p2panda/pull/284) `rs`
+- Introduce new application schema id format [#292](https://github.com/p2panda/p2panda/pull/292) `rs`
 
 ## Fixed